fix: refactor update.sh script to include project_name parameter
project name parameter will now be used in docker-compose command in
order to setup dd_platform docker installation.

update.sh scripts have been updated in order to check the presence of
this mandatory parameter.

It'll also check if a docker container associated with this project name
is already running on the docker environment.